Extract view model to the separate file
